moved monkey patches to __init__.py
